+++ src/components/HomePage/AccountCard/AccountCard.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { Card } from 'decentraland-ui'
+import { Transaction, TransactionStatus } from '../HomePage.types'
 import { AccountCardHeader } from './AccountCardHeader'
-import { Transaction, TransactionStatus, Props } from './AccountCard.types'
+import { Props } from './AccountCard.types'
 import { AccountTransactions } from './AccountTransactions'
 import './AccountCard.css'
 
@@ -9,7 +10,7 @@ const AccountCard = ({
   type,
   title,
   onFetchAmount,
-  onFetchTransactions
+  onFetchTransactions,
 }: Props) => {
   const [amount, setAmount] = useState(0)
   const [transactions, setTransactions]: [Transaction[], any] = useState([])
