+++ src/components/HomePage/AccountCard/AccountCard.types.tsx
@@ -1,24 +1,4 @@
-export enum NetworkType {
-  MATIC = 'matic',
-  ETHEREUM = 'ethereum',
-}
-
-export enum TransactionType {
-  DEPOSIT = 'deposit',
-  WITHDRAWAL = 'withdrawal',
-}
-
-export enum TransactionStatus {
-  CONFIRMED = 'confirmed',
-  PENDING = 'pending',
-}
-
-export type Transaction = {
-  type: TransactionType
-  status: TransactionStatus
-  amount: number
-  description: string
-}
+import { NetworkType, Transaction } from '../HomePage.types'
 
 export type Props = {
   type: NetworkType
