+++ src/components/HomePage/AccountCard/AccountTransactions/AccountTransaction/AccountTransaction.types.tsx
@@ -1,3 +1,3 @@
-import { Transaction } from '../../AccountCard.types'
+import { Transaction } from '../../../HomePage.types'
 
 export type Props = Transaction
