+++ src/components/HomePage/HomePage.tsx
@@ -4,10 +4,10 @@ import { Footer, Navbar } from 'decentraland-dapps/dist/containers'
 import { AccountHeader } from './AccountHeader'
 import {
   NetworkType,
+  Props,
   TransactionStatus,
-  TransactionType
-} from './AccountCard/AccountCard.types'
-import { Props } from './HomePage.types'
+  TransactionType,
+} from './HomePage.types'
 import './HomePage.css'
 import { AccountCard, AccountCardContainer } from './AccountCard'
 
@@ -27,20 +27,20 @@ const HomePage = (_props: Props) => {
                 type: TransactionType.DEPOSIT,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
+                amount: 3000,
               },
               {
                 type: TransactionType.WITHDRAWAL,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
+                amount: 3000,
               },
               {
                 type: TransactionType.DEPOSIT,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
-              }
+                amount: 3000,
+              },
             ]}
           />
           <AccountCard
@@ -52,20 +52,20 @@ const HomePage = (_props: Props) => {
                 type: TransactionType.DEPOSIT,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
+                amount: 3000,
               },
               {
                 type: TransactionType.WITHDRAWAL,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
+                amount: 3000,
               },
               {
                 type: TransactionType.DEPOSIT,
                 description: 'Added tokens via transak',
                 status: TransactionStatus.CONFIRMED,
-                amount: 3000
-              }
+                amount: 3000,
+              },
             ]}
           />
         </AccountCardContainer>
