+++ src/components/HomePage/HomePage.types.tsx
@@ -1,1 +1,23 @@
 export type Props = {}
+
+export enum NetworkType {
+  MATIC = 'matic',
+  ETHEREUM = 'ethereum',
+}
+
+export enum TransactionType {
+  DEPOSIT = 'deposit',
+  WITHDRAWAL = 'withdrawal',
+}
+
+export enum TransactionStatus {
+  CONFIRMED = 'confirmed',
+  PENDING = 'pending',
+}
+
+export type Transaction = {
+  type: TransactionType
+  status: TransactionStatus
+  amount: number
+  description: string
+}
